# plan-ai/backend/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ApiResponse_null_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"anyOf": [],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"TrelloManualConnectRequest": {
				"properties": {
					"apiKey": {
						"type": "string"
					},
					"token": {
						"type": "string"
					}
				},
				"required": [
					"apiKey",
					"token"
				],
				"type": "object",
				"additionalProperties": false
			},
			"TrelloSummaryResponse": {
				"properties": {
					"totalBoards": {
						"type": "number",
						"format": "double"
					},
					"latestBoards": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"totalBoards",
					"latestBoards"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_TrelloSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TrelloSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"_36_Enums.ProjectStatus": {
				"type": "string",
				"enum": [
					"ACTIVE",
					"COMPLETED",
					"ARCHIVED"
				]
			},
			"ProjectStatus": {
				"$ref": "#/components/schemas/_36_Enums.ProjectStatus"
			},
			"TsoaJsonObject": {
				"anyOf": [
					{
						"properties": {},
						"additionalProperties": {},
						"type": "object"
					},
					{
						"items": {},
						"type": "array"
					},
					{
						"type": "string"
					},
					{
						"type": "number",
						"format": "double"
					},
					{
						"type": "boolean"
					}
				],
				"nullable": true
			},
			"ProjectResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string",
						"nullable": true
					},
					"status": {
						"$ref": "#/components/schemas/ProjectStatus"
					},
					"startedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"endedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"metadata": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TsoaJsonObject"
							}
						],
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"status",
					"startedAt",
					"endedAt",
					"metadata",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProjectListResponse": {
				"properties": {
					"projects": {
						"items": {
							"$ref": "#/components/schemas/ProjectResponse"
						},
						"type": "array"
					},
					"total": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"projects",
					"total"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_ProjectListResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ProjectListResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"ApiResponse_ProjectResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/ProjectResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"CreateProjectRequest": {
				"properties": {
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"status": {
						"$ref": "#/components/schemas/ProjectStatus"
					},
					"startedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"metadata": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TsoaJsonObject"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"title"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ApiResponse__success-boolean__": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"success": {
								"type": "boolean"
							}
						},
						"required": [
							"success"
						],
						"type": "object",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"LinearManualConnectRequest": {
				"properties": {
					"apiKey": {
						"type": "string"
					}
				},
				"required": [
					"apiKey"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LinearSummaryResponse": {
				"properties": {
					"totalIssues": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"totalProjects": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"latestTeams": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"totalIssues",
					"totalProjects",
					"latestTeams"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_LinearSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/LinearSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
			"JiraAuthorizationResponse": {
				"properties": {
					"authorizationUrl": {
						"type": "string"
					}
				},
				"required": [
					"authorizationUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_JiraAuthorizationResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/JiraAuthorizationResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"JiraManualConnectRequest": {
				"properties": {
					"siteUrl": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"apiToken": {
						"type": "string"
					}
				},
				"required": [
					"siteUrl",
					"email",
					"apiToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"JiraSummaryResponse": {
				"properties": {
					"totalIssues": {
						"type": "string",
						"nullable": true
					},
					"totalProjects": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"latestBoards": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"totalIssues",
					"totalProjects",
					"latestBoards"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_JiraSummaryResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/JiraSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"JiraProjectItem": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"key": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"key"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_JiraProjectItem-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/JiraProjectItem"
						},
						"type": "array",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"SetDefaultProjectRequest": {
				"properties": {
					"projectId": {
						"type": "string"
					}
				},
				"required": [
					"projectId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"_36_Enums.IntegrationProvider": {
				"type": "string",
				"enum": [
					"JIRA",
					"LINEAR",
					"GITHUB",
					"GOOGLE_DRIVE",
					"TRELLO"
				]
			},
			"IntegrationProvider": {
				"$ref": "#/components/schemas/_36_Enums.IntegrationProvider"
			},
			"_36_Enums.IntegrationStatus": {
				"type": "string",
				"enum": [
					"CONNECTED",
					"DISCONNECTED",
					"ERROR"
				]
			},
			"IntegrationStatus": {
				"$ref": "#/components/schemas/_36_Enums.IntegrationStatus"
			},
			"IntegrationSummaryResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"provider": {
						"$ref": "#/components/schemas/IntegrationProvider"
					},
					"status": {
						"$ref": "#/components/schemas/IntegrationStatus"
					},
					"accountId": {
						"type": "string",
						"nullable": true
					},
					"accountName": {
						"type": "string",
						"nullable": true
					},
					"metadata": {
						"allOf": [
							{
								"$ref": "#/components/schemas/TsoaJsonObject"
							}
						],
						"nullable": true
					},
					"scope": {
						"type": "string",
						"nullable": true
					},
					"expiresAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"hasRefreshToken": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"provider",
					"status",
					"accountId",
					"accountName",
					"metadata",
					"scope",
					"expiresAt",
					"createdAt",
					"updatedAt",
					"hasRefreshToken"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_IntegrationSummaryResponse-Array_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/IntegrationSummaryResponse"
						},
						"type": "array",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"ApiResponse_IntegrationSummaryResponse-or-null_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IntegrationSummaryResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "BUSL-1.1"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/trello/manual-connect": {
			"post": {
				"operationId": "ManualConnect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/TrelloManualConnectRequest"
							}
						}
					}
				}
			}
		},
		"/api/trello/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_TrelloSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/projects": {
			"get": {
				"operationId": "ListSessions",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ProjectListResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "page",
						"required": false,
						"schema": {
							"default": 1,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "status",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/ProjectStatus"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateSession",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_ProjectResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Projects"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateProjectRequest"
							}
						}
					}
				}
			}
		},
...
		"/api/linear/manual-connect": {
			"post": {
				"operationId": "ManualConnect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					},
					"400": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/LinearManualConnectRequest"
							}
						}
					}
				}
			}
		},
		"/api/linear/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_LinearSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
...
		"/api/jira/auth": {
			"get": {
				"operationId": "GetAuthorizationUrl",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_JiraAuthorizationResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/jira/callback": {
			"get": {
				"operationId": "HandleCallback",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "error",
						"required": false,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "error_description",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/jira/manual-connect": {
			"post": {
				"operationId": "ManualConnect",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__success-boolean__"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/JiraManualConnectRequest"
							}
						}
					}
				}
			}
		},
		"/api/jira/summary": {
			"get": {
				"operationId": "GetSummary",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_JiraSummaryResponse_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
		"/api/jira/projects": {
			"get": {
				"operationId": "GetProjects",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_JiraProjectItem-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
		"/api/jira/default-project": {
			"post": {
				"operationId": "SetDefaultProject",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SetDefaultProjectRequest"
							}
						}
					}
				}
			}
		},
		"/api/integrations": {
			"get": {
				"operationId": "ListIntegrations",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_IntegrationSummaryResponse-Array_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": []
			}
		},
		"/api/integrations/{provider}": {
			"get": {
				"operationId": "GetIntegration",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_IntegrationSummaryResponse-or-null_"
								}
							}
						}
					}
				},
				"tags": [
					"Integrations"
				],
				"security": [
					{
						"ClientLevel": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "provider",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}